Instructor notes for teaching git via collaboration on GitHub.
The focus of the lesson is to have students contribute to a repo on GitHub in
which we are building up a list of the things they are most interested in
learning at the boot camp. Requires everyone have a good internet connection
and that at least some of the participants sign up for GitHub.