make/Makefile: Update to use POSIX syntax
The old version had a number of problems.  For example, if you didn't
have `.` in your PATH:

  $ make
  make_plots.py kepler_planets.csv exoplanet.eu_catalog.csv
  make: make_plots.py: Command not found
  make: *** [*.png] Error 127

which I fixed by adding a PYTHON macro.  As a side benefit, this also
allows you to override the default version of Python from the command
line.  For example, I have Python 3.2 as my system Python, but only
have Matplotlib installed for Python 2.7.  With the new Makefile, I
can run:

  $ make PYTHON=python2.7

Another command problem was the lack of `-f` in the `clean` rule,
which lead to:

  $ make clean
  rm *.png exoplanet_figures.html
  rm: cannot remove ‘*.png’: No such file or directory
  rm: cannot remove ‘exoplanet_figures.html’: No such file or directory
  make: *** [clean] Error 1

if you tried to clean an already- or partially-clean directory.

The use of `*.png` as a target and prerequisite was also troublesome
(besides being a GNU extension [1]), because it made it impossible for
Make to properly track dependencies.  For example:

  $ make
  ./make_plots.py kepler_planets.csv exoplanet.eu_catalog.csv
  ./make_html.py exoplanet_figures.html *.png
  $ touch kepler_planets.csv
  $ make
  ./make_plots.py kepler_planets.csv exoplanet.eu_catalog.csv
  ./make_html.py exoplanet_figures.html kepler_planets.png exoplanet.eu_catalog.png

Note that it is rebuilding kepler_planets.png (which is good, because
I touched it).  It is *also* rebuilding exoplanet.eu_catalog.png
(which is wrong, because the source hasn't changed).  With the new
Makefile we get the appropriate behaviour:

  $ make
  python make_plots.py kepler_planets.csv
  python make_plots.py exoplanet.eu_catalog.csv
  python make_html.py exoplanet_figures.html kepler_planets.png exoplanet.eu_catalog.png
  $ touch kepler_planets.csv
  $ make
  python2.7 make_plots.py kepler_planets.csv
  python2.7 make_html.py exoplanet_figures.html kepler_planets.png exoplanet.eu_catalog.png

I also added a number of comments referencing the POSIX make
specification, since that seems like a good least-common denominator.
If we decide that Software Carpentry can drop support for non-GNU
implementations, these links should be updated to point into the GNU
Make Manual.  It would also be a good idea to replace the more obscure
double-suffix inference rule with a pattern rule [2].  It might also
be a good idea to use the wildcard funtion to populate CSV_SOURCE [3].

The comments make the Makefile a bit more difficult for experienced
users to read by spreading out the non-comment lines.  However, I
think the comments will help new Make users (who haven't seen the
syntax before) get a feel for what a line is doing.

[1]: http://www.gnu.org/software/make/manual/html_node/Wildcards.html
[2]: http://www.gnu.org/software/make/manual/html_node/Pattern-Intro.html
[3]: http://www.gnu.org/software/make/manual/html_node/Wildcard-Function.html