make: Remove this whole section
We don't teach compiled languages, so the need to teach build systems
is not very strong.  I also use Make to help analyze my data, but
nobody has had the time or inclination to teach Make in a boot camp
yet (as far as I know).  Let's just remove this whole section.