Avoid current Canopy brakage
IPython in the current version of Canopy has a __version__
of '1.0.dev' and int('.dev') breaks the version number checking.
This leads to the error:

check for IPython Python package (IPython) failed:
  unparsable 'dev' in version 1.0.dev of IPython Python package (IPython), (need >= 0.13)
  For instructions on installing an up-to-date version, see
  http://ipython.org/install.html

To avoid this we can do s/'.dev'/'.0'/ for python module versions
before doing the version number tests making sure to handle the
version = None case.