Improve type discussion somewhat
Add some examples to the initial discussion of types.

Also, talking about "fear and loathing" is exactly a good sell for a language intro, so rework that bit.

Also, the concepts of int and float haven't yet been introduced. So explain them a bit. It might be saner to drop the "let's make it a float" part but I didn't do that.