Remove discussion of `is` in "Write Code for People"
It's rarely-used (I write a lot of python and basically
never use it) and really special-case-y. It's mostly
useful for `if foo is None` but we don't ever cover
`None`, so there's that. And `is` doesn't come up in
any of the rest of the curriculum.

Anyhow, we never cover whitespace in the current
version, so we'll need space from somewhere.